fixes #43: strip CD from damage before trying to math it
