+++ fallout-2d20-gui/src/screens/stats.rs
@@ -1,5 +1,6 @@
 use imgui::Ui;
 use sdl2::video::Window;
+use crate::theme::BAR_HEIGHT;
 use crate::{AppScreen, render_placeholder};
 use crate::screens::special::SpecialState;
 use crate::screens::skills::{ SkillsState, SKILLS };
@@ -189,10 +190,10 @@ pub fn render_stats(
     state: &mut ComputedStats,
 ) {
     let (win_w, win_h) = window.size();
-    let w = (win_w as f32 * 0.65).min(960.0);
-    let h = win_h as f32 * 0.85;
-    let bar_h = crate::BAR_HEIGHT;
+    let bar_h = BAR_HEIGHT;
     let content_h = win_h as f32 - bar_h;
+    let w = (win_w as f32 * 0.85).min(1100.0);
+    let h = content_h * 0.92;
 
     let Some(_tok) = ui.window("##stats")
         .title_bar(false)
